+++ src/main/java/kr/re/kitri/springdemo/dao/ArticleDao.java
@@ -38,7 +38,7 @@ public class ArticleDao {
         return "글쓰기가 완료되었습니다. from dao";
     }
 
-    public Article selectArticleByArticleId(int articleId) {
+    public Article selectArticleByArticleId(long articleId) {
         return new Article(articleId, "aaa", "aaa", "aaaa", null, 10);
     }
 }
